⚡️ [refactoring] 메인모달관련해서 여러 파일로 분리하기 #11
