+++ src/utils/index.js
@@ -8,3 +8,4 @@ export * from './attr';
 export * from './getNode';
 export * from './typeOf';
 export * from './typeError';
+export * from './getCookie';
